[AND-844] Move RedeemRewardScreen to unit_exchange_flow package
