+++ app-games/src/main/java/com/aptoide/android/aptoidegames/toolbar/AppGamesToolBar.kt
@@ -41,7 +41,7 @@ import com.aptoide.android.aptoidegames.drawables.icons.getProfileNoAccountIcon
 import com.aptoide.android.aptoidegames.notifications.NotificationsPermissionRequester
 import com.aptoide.android.aptoidegames.play_and_earn.presentation.components.PlayAndEarnTopBarBadge
 import com.aptoide.android.aptoidegames.play_and_earn.presentation.sign_in.rememberUserInfo
-import com.aptoide.android.aptoidegames.play_and_earn.presentation.units.redeemRewardRoute
+import com.aptoide.android.aptoidegames.play_and_earn.presentation.unit_exchange_flow.redeemRewardRoute
 import com.aptoide.android.aptoidegames.settings.settingsRoute
 import com.aptoide.android.aptoidegames.theme.Palette
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
